Separated profile and chat
